requirements.txt
numpy>=1.24
torch>=2.1
transformers>=4.40
datasets>=2.18
scikit-learn>=1.2
tqdm>=4.66
pyyaml>=6.0
pandas>=2.0
safetensors>=0.4.0
pyarrow>=10.0.0
